fix(cloneservergrouptask): undo the move of CloneServerGroupTask
This change undoes the move of CloneServerGroupTask to a different package
(this was introduced in #2815).

Today, tasks can not be moved once they have been created because Orca will
be unable to deserialize the tasks already in the queue created under a
different package.
